requirements.txt
alembic==0.9.9
arrow==0.12.1
asn1crypto==0.24.0
astroid==1.6.5
autopep8==1.3.5
blinker==1.4
cffi==1.11.5
click==6.7
cryptography==1.9
dateparser==0.7.0
dominate==2.3.1
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Moment==0.6.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
guess-language-spirit==0.5.3
idna==2.7
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
jwt==0.5.3
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
migrate==0.3.8
moment==0.8.2
mysqlclient==1.3.12
pycodestyle==2.4.0
pycparser==2.18
PyJWT==1.6.4
pylint==1.9.2
python-dateutil==2.7.3
python-editor==1.0.3
pytz==2018.5
regex==2018.6.21
six==1.11.0
SQLAlchemy==1.2.7
times==0.7
tzlocal==1.5.1
visitor==0.1.3
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.1
